Split resources
