+++ app/build.gradle
@@ -52,9 +52,11 @@ android {
     sourceSets {
         full {
             java.srcDirs = ['src/full/java']
+            res.srcDirs = ['src/full/java/res']
         }
         demo {
             java.srcDirs = ['src/demo/java']
+            res.srcDirs = ['src/demo/java/res']
         }
     }
 }
