+++ app/src/demo/java/res/drawable/ic_man_1.xml
