+++ app/src/demo/java/res/drawable/ic_man_2.xml
