+++ app/src/demo/java/res/drawable/ic_man_3.xml
